Move ConstantNotUsedInspectionTests for QuickFixes into relevant files under RD.QuickFixes
