+++ RubberduckTests/Inspections/ConstantNotUsedInspectionTests.cs
@@ -183,60 +183,6 @@ End Sub";
 
         [TestMethod]
         [TestCategory("Inspections")]
-        public void ConstantNotUsed_QuickFixWorks()
-        {
-            const string inputCode =
-@"Public Sub Foo()
-Const const1 As Integer = 9
-End Sub";
-
-            const string expectedCode =
-@"Public Sub Foo()
-End Sub";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new ConstantNotUsedInspection(state);
-            var inspectionResults = inspection.GetInspectionResults();
-
-            new RemoveUnusedDeclarationQuickFix(state).Fix(inspectionResults.First());
-
-            var rewriter = state.GetRewriter(component);
-            var rewrittenCode = rewriter.GetText();
-            Assert.AreEqual(expectedCode, rewrittenCode);
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void ConstantNotUsed_IgnoreQuickFixWorks()
-        {
-            const string inputCode =
-@"Public Sub Foo()
-    Const const1 As Integer = 9
-End Sub";
-
-            const string expectedCode =
-@"Public Sub Foo()
-'@Ignore ConstantNotUsed
-    Const const1 As Integer = 9
-End Sub";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new ConstantNotUsedInspection(state);
-            var inspectionResults = inspection.GetInspectionResults();
-            
-            new IgnoreOnceQuickFix(state, new[] {inspection}).Fix(inspectionResults.First());
-
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
         public void InspectionType()
         {
             var inspection = new ConstantNotUsedInspection(null);
